+++ mithril-common/src/entities/epoch_settings.rs
@@ -1,4 +1,4 @@
-use crate::entities::{Epoch, ProtocolParameters};
+use crate::entities::{CardanoTransactionsSigningConfig, Epoch, ProtocolParameters};
 
 use super::Signer;
 
@@ -19,4 +19,10 @@ pub struct EpochSettings {
 
     /// Signers that will be able to sign on the next epoch
     pub next_signers: Vec<Signer>,
+
+    /// Cardano transactions signing configuration for the current epoch
+    pub cardano_transactions_signing_config: Option<CardanoTransactionsSigningConfig>,
+
+    /// Cardano transactions signing configuration for the next epoch
+    pub next_cardano_transactions_signing_config: Option<CardanoTransactionsSigningConfig>,
 }