+++ mithril-common/src/messages/epoch_settings.rs
@@ -1,4 +1,4 @@
-use crate::entities::{Epoch, ProtocolParameters};
+use crate::entities::{CardanoTransactionsSigningConfig, Epoch, ProtocolParameters};
 use crate::messages::SignerMessagePart;
 use serde::{Deserialize, Serialize};
 
@@ -21,6 +21,14 @@ pub struct EpochSettingsMessage {
 
     /// Signers that will be able to sign on the next epoch
     pub next_signers: Vec<SignerMessagePart>,
+
+    /// Cardano transactions signing configuration for the current epoch
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub cardano_transactions_signing_config: Option<CardanoTransactionsSigningConfig>,
+
+    /// Cardano transactions signing configuration for the next epoch
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub next_cardano_transactions_signing_config: Option<CardanoTransactionsSigningConfig>,
 }
 
 impl EpochSettingsMessage {
@@ -41,6 +49,8 @@ impl EpochSettingsMessage {
                 },
                 current_signers: [SignerMessagePart::dummy()].to_vec(),
                 next_signers: [SignerMessagePart::dummy()].to_vec(),
+                cardano_transactions_signing_config: Some(CardanoTransactionsSigningConfig::dummy()),
+                next_cardano_transactions_signing_config: Some(CardanoTransactionsSigningConfig::dummy()),
             }
         }
     }
@@ -48,6 +58,8 @@ impl EpochSettingsMessage {
 
 #[cfg(test)]
 mod tests {
+
+    use crate::entities::BlockNumber;
 
     use super::*;
 
@@ -68,12 +80,21 @@ mod tests {
             "verification_key_signature":"signature_456",
             "operational_certificate":"certificate_456",
             "kes_period":45
-        }]
+        }],
+        "cardano_transactions_signing_config": {
+            "security_parameter": 70,
+            "step": 20
+        },
+        "next_cardano_transactions_signing_config": {
+            "security_parameter": 50,
+            "step": 10
+        }
         
         }"#;
 
+    // Supported structure until OpenAPI version 0.1.28.
     #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
-    pub struct EpochSettingsMessagePreviousVersion {
+    pub struct EpochSettingsMessageUntilV0_1_28 {
         /// Current Epoch
         pub epoch: Epoch,
 
@@ -86,8 +107,29 @@ mod tests {
         pub next_protocol_parameters: ProtocolParameters,
     }
 
-    fn golden_previous_message() -> EpochSettingsMessagePreviousVersion {
-        EpochSettingsMessagePreviousVersion {
+    // Supported structure until OpenAPI version 0.1.29.
+    #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
+    pub struct EpochSettingsMessageUntilV0_1_29 {
+        /// Current Epoch
+        pub epoch: Epoch,
+
+        /// Current Protocol parameters
+        #[serde(rename = "protocol")]
+        pub protocol_parameters: ProtocolParameters,
+
+        /// Next Protocol parameters
+        #[serde(rename = "next_protocol")]
+        pub next_protocol_parameters: ProtocolParameters,
+
+        /// Current Signers
+        pub current_signers: Vec<SignerMessagePart>,
+
+        /// Signers that will be able to sign on the next epoch
+        pub next_signers: Vec<SignerMessagePart>,
+    }
+
+    fn golden_message_until_open_api_0_1_28() -> EpochSettingsMessageUntilV0_1_28 {
+        EpochSettingsMessageUntilV0_1_28 {
             epoch: Epoch(10),
             protocol_parameters: ProtocolParameters {
                 k: 5,
@@ -99,6 +141,36 @@ mod tests {
                 m: 1000,
                 phi_f: 0.65,
             },
+        }
+    }
+
+    fn golden_message_until_open_api_0_1_29() -> EpochSettingsMessageUntilV0_1_29 {
+        EpochSettingsMessageUntilV0_1_29 {
+            epoch: Epoch(10),
+            protocol_parameters: ProtocolParameters {
+                k: 5,
+                m: 100,
+                phi_f: 0.65,
+            },
+            next_protocol_parameters: ProtocolParameters {
+                k: 50,
+                m: 1000,
+                phi_f: 0.65,
+            },
+            current_signers: vec![SignerMessagePart {
+                party_id: "123".to_string(),
+                verification_key: "key_123".to_string(),
+                verification_key_signature: Some("signature_123".to_string()),
+                operational_certificate: Some("certificate_123".to_string()),
+                kes_period: Some(12),
+            }],
+            next_signers: vec![SignerMessagePart {
+                party_id: "456".to_string(),
+                verification_key: "key_456".to_string(),
+                verification_key_signature: Some("signature_456".to_string()),
+                operational_certificate: Some("certificate_456".to_string()),
+                kes_period: Some(45),
+            }],
         }
     }
 
@@ -129,18 +201,37 @@ mod tests {
                 operational_certificate: Some("certificate_456".to_string()),
                 kes_period: Some(45),
             }],
+            cardano_transactions_signing_config: Some(CardanoTransactionsSigningConfig {
+                security_parameter: BlockNumber(70),
+                step: BlockNumber(20),
+            }),
+            next_cardano_transactions_signing_config: Some(CardanoTransactionsSigningConfig {
+                security_parameter: BlockNumber(50),
+                step: BlockNumber(10),
+            }),
         }
     }
 
-    // Test the backward compatibility with previous structure.
+    // Test the backward compatibility with the structure supported until OpenAPI version 0.1.28.
     #[test]
-    fn test_actual_json_deserialized_into_previous_message() {
+    fn test_actual_json_deserialized_into_message_supported_until_open_api_0_1_28() {
         let json = ACTUAL_JSON;
-        let message: EpochSettingsMessagePreviousVersion = serde_json::from_str(json).expect(
-            "This JSON is expected to be successfully parsed into a EpochSettingsMessagePreviousVersion instance.",
+        let message: EpochSettingsMessageUntilV0_1_28 = serde_json::from_str(json).expect(
+            "This JSON is expected to be successfully parsed into a EpochSettingsMessageUntilVersion0_1_28 instance.",
         );
 
-        assert_eq!(golden_previous_message(), message);
+        assert_eq!(golden_message_until_open_api_0_1_28(), message);
+    }
+
+    // Test the backward compatibility with the structure supported until OpenAPI version 0.1.29.
+    #[test]
+    fn test_actual_json_deserialized_into_message_supported_until_open_api_0_1_29() {
+        let json = ACTUAL_JSON;
+        let message: EpochSettingsMessageUntilV0_1_29 = serde_json::from_str(json).expect(
+            "This JSON is expected to be successfully parsed into a EpochSettingsMessageUntilVersion0_1_29 instance.",
+        );
+
+        assert_eq!(golden_message_until_open_api_0_1_29(), message);
     }
 
     // Test the compatibility with current structure.