+++ mithril-signer/src/message_adapters/from_epoch_settings.rs
@@ -19,6 +19,9 @@ impl TryFromMessageAdapter<EpochSettingsMessage, EpochSettings> for FromEpochSet
                 .with_context(|| "'FromMessageAdapter' can not convert the current signers")?,
             next_signers: SignerMessagePart::try_into_signers(message.next_signers)
                 .with_context(|| "'FromMessageAdapter' can not convert the next signers")?,
+            cardano_transactions_signing_config: message.cardano_transactions_signing_config,
+            next_cardano_transactions_signing_config: message
+                .next_cardano_transactions_signing_config,
         };
         Ok(epoch_settings)
     }