+++ mithril-signer/src/services/epoch_service.rs
@@ -2,6 +2,7 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use mithril_common::crypto_helper::ProtocolInitializer;
+use mithril_common::entities::CardanoTransactionsSigningConfig;
 use mithril_common::entities::Epoch;
 use mithril_common::entities::PartyId;
 use mithril_common::entities::ProtocolParameters;
@@ -49,6 +50,16 @@ pub trait EpochService: Sync + Send {
     /// Get signers with stake for the next epoch
     async fn next_signers_with_stake(&self) -> StdResult<Vec<SignerWithStake>>;
 
+    /// Get the cardano transactions signing configuration for the current epoch
+    fn cardano_transactions_signing_config(
+        &self,
+    ) -> StdResult<&Option<CardanoTransactionsSigningConfig>>;
+
+    /// Get the cardano transactions signing configuration for the next epoch
+    fn next_cardano_transactions_signing_config(
+        &self,
+    ) -> StdResult<&Option<CardanoTransactionsSigningConfig>>;
+
     /// Check if a signer is included in the current stake distribution
     fn is_signer_included_in_current_stake_distribution(
         &self,
@@ -62,6 +73,8 @@ struct EpochData {
     next_protocol_parameters: ProtocolParameters,
     current_signers: Vec<Signer>,
     next_signers: Vec<Signer>,
+    cardano_transactions_signing_config: Option<CardanoTransactionsSigningConfig>,
+    next_cardano_transactions_signing_config: Option<CardanoTransactionsSigningConfig>,
 }
 
 /// Implementation of the [epoch service][EpochService].
@@ -138,6 +151,9 @@ impl EpochService for MithrilEpochService {
             next_protocol_parameters: epoch_settings.next_protocol_parameters,
             current_signers: epoch_settings.current_signers,
             next_signers: epoch_settings.next_signers,
+            cardano_transactions_signing_config: epoch_settings.cardano_transactions_signing_config,
+            next_cardano_transactions_signing_config: epoch_settings
+                .next_cardano_transactions_signing_config,
         });
 
         Ok(())
@@ -175,6 +191,18 @@ impl EpochService for MithrilEpochService {
             self.next_signers()?,
         )
         .await
+    }
+
+    fn cardano_transactions_signing_config(
+        &self,
+    ) -> StdResult<&Option<CardanoTransactionsSigningConfig>> {
+        Ok(&self.unwrap_data()?.cardano_transactions_signing_config)
+    }
+
+    fn next_cardano_transactions_signing_config(
+        &self,
+    ) -> StdResult<&Option<CardanoTransactionsSigningConfig>> {
+        Ok(&self.unwrap_data()?.next_cardano_transactions_signing_config)
     }
 
     fn is_signer_included_in_current_stake_distribution(
@@ -302,6 +330,8 @@ mod tests {
         assert!(service.next_signers().is_err());
         assert!(service.current_signers_with_stake().await.is_err());
         assert!(service.next_signers_with_stake().await.is_err());
+        assert!(service.cardano_transactions_signing_config().is_err());
+        assert!(service.next_cardano_transactions_signing_config().is_err());
     }
 
     #[test]
@@ -349,6 +379,17 @@ mod tests {
         assert_eq!(
             epoch_settings.next_protocol_parameters,
             *service.next_protocol_parameters().unwrap()
+        );
+
+        // Check cardano_transactions_signing_config
+        assert_eq!(
+            epoch_settings.cardano_transactions_signing_config,
+            *service.cardano_transactions_signing_config().unwrap()
+        );
+        // Check next_cardano_transactions_signing_config
+        assert_eq!(
+            epoch_settings.next_cardano_transactions_signing_config,
+            *service.next_cardano_transactions_signing_config().unwrap()
         );
     }